% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{display.config}
\alias{display.config}
\title{Set the chart config metadata for current dataframe (set by \code{display}) which needs to be visualized.}
\usage{
display.config(
  commId,
  lastCommId = NULL,
  binsNumber = 10,
  category = "table",
  keys = NULL,
  values = NULL,
  series = NULL,
  aggregation = NULL,
  column = NULL
)
}
\arguments{
\item{commId}{the id used to identify whether the API call from synapse notebook js client.}

\item{lastCommId}{same with \code{id} parameter, but the previous value.}

\item{binsNumber}{bins number for rendering histogram, default is 10.}

\item{category}{the chart category as \code{bar}, \code{line}, default is \code{table}.}

\item{keys}{the column names which useds to render x-axis.}

\item{values}{the column names which used to render y-axis.}

\item{series}{the column which used to render the chart series}

\item{aggregation}{the aggregation operation type: \code{sum}, \code{avg}, \code{min}, \code{max}, \code{count}.}

\item{column}{will deperated: the column name used to calculate the statistic info,
as the column type, unique values, missing values, etc.}
}
\description{
Set the chart config metadata for current dataframe (set by \code{display}) which needs to be visualized.
}
