% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{displayHTML}
\alias{displayHTML}
\title{notebook.R: Utility for visualization
Construct an specific html fragment to synapse notebook front-end for rendering
based on user-input html content.}
\usage{
displayHTML(content)
}
\arguments{
\item{content}{html content which user want to render}
}
\value{
No return value, print the \code{content} to mimic the render behavior when used in azure synapse runtime.
}
\description{
notebook.R: Utility for visualization
Construct an specific html fragment to synapse notebook front-end for rendering
based on user-input html content.
}
\examples{
displayHTML('<b>Hello world!</b>')

}
