% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{mssparkutils.credentials.getConnectionStringOrCreds}
\alias{mssparkutils.credentials.getConnectionStringOrCreds}
\title{Take linked service name as input and return connection string or credentials
depending on the configuration of the linked service.}
\usage{
mssparkutils.credentials.getConnectionStringOrCreds(linkedService)
}
\arguments{
\item{linkedService}{Linked service name.}
}
\value{
A empty string used to mimic credentials returned by azure synapse runtime for \code{linkedService}.
}
\description{
Take linked service name as input and return connection string or credentials
depending on the configuration of the linked service.
}
\examples{
mssparkutils.credentials.getConnectionStringOrCreds('AzureDataLakeStorage1')

}
