% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{mssparkutils.credentials.isValidToken}
\alias{mssparkutils.credentials.isValidToken}
\title{Returns true if the input token is valid (i.e, hasn't expired).}
\usage{
mssparkutils.credentials.isValidToken(token)
}
\arguments{
\item{token}{token to validate.}
}
\value{
FALSE to mimic the result if \code{token} is invalid.
}
\description{
Returns true if the input token is valid (i.e, hasn't expired).
}
\examples{
mssparkutils.credentials.isValidToken('dummyToken')

}
