% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{mssparkutils.env.getJobId}
\alias{mssparkutils.env.getJobId}
\title{Get job Id.}
\usage{
mssparkutils.env.getJobId()
}
\value{
A empty string used to mimic the id of spark job been submitted to azure synapse runtime.
}
\description{
Get job Id.
}
\examples{
mssparkutils.env.getJobId()

}
