% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{mssparkutils.fs.append}
\alias{mssparkutils.fs.append}
\title{Append the given String to a file, encoded in UTF-8.}
\usage{
mssparkutils.fs.append(file, content, createFileIfNotExists = FALSE)
}
\arguments{
\item{file}{FileSystem URI}

\item{content}{Content needs to be append to file, encoded in System default charset.}

\item{createFileIfNotExists}{If set to true, will firstly try to create file if not exists.}
}
\value{
FALSE to mimic the result if \code{file} content append fail.
}
\description{
Append the given String to a file, encoded in UTF-8.
}
\examples{
mssparkutils.fs.append("/tmp/my-file", "Hello world!")
mssparkutils.fs.append("/tmp/my-file", "Hello world!", TRUE)

}
