% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{mssparkutils.fs.fastcp}
\alias{mssparkutils.fs.fastcp}
\title{Copies a file or directory via azcopy, possibly across FileSystems.}
\usage{
mssparkutils.fs.fastcp(from, to, recurse = TRUE, extraConfigs = NULL)
}
\arguments{
\item{from}{FileSystem URI of the source file or directory}

\item{to}{FileSystem URI of the destination file or directory}

\item{recurse}{if TRUE, all files and directories will be recursively copied}

\item{extraConfigs}{extra configs for azcopy, includes flags, timeout, aadToken, sourceLinkedService, destinationLinkedService}
}
\value{
TRUE if all files were successfully copied
}
\description{
Copies a file or directory via azcopy, possibly across FileSystems.
}
\examples{
\dontrun{
mssparkutils.fs.fastcp("file:/tmp/my-folder/a", "adls://xxx/tmp/b")
}

}
