% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{mssparkutils.fs.getMountPath}
\alias{mssparkutils.fs.getMountPath}
\title{Gets the local path of the mount point.}
\usage{
mssparkutils.fs.getMountPath(mountPoint, scope = "job")
}
\arguments{
\item{mountPoint}{The directory that was previously mounted.}

\item{scope}{Mount point level, job or workspace, default is job.}
}
\value{
Empty string to mimic the local mounted path related to \code{mountPoint}.
}
\description{
Gets the local path of the mount point.
}
\examples{
mssparkutils.fs.getMountPath("/mnt")
mssparkutils.fs.getMountPath("/mnt", "job")

}
