% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{mssparkutils.fs.mkdirs}
\alias{mssparkutils.fs.mkdirs}
\title{Creates the given directory if it does not exist, also creating any necessary parent * directories.}
\usage{
mssparkutils.fs.mkdirs(dir)
}
\arguments{
\item{dir}{FileSystem URI}
}
\value{
FALSE to mimic the result if \code{dir} creation fail.
}
\description{
Creates the given directory if it does not exist, also creating any necessary parent * directories.
}
\examples{
mssparkutils.fs.mkdirs("/tmp/a/b/c")

}
