% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{mssparkutils.fs.mount}
\alias{mssparkutils.fs.mount}
\title{Attach remote storage (Blob, Gen2, Azure File Share) to all working nodes (driver node and worker nodes)}
\usage{
mssparkutils.fs.mount(source, mountPoint, extraConfigs = NULL)
}
\arguments{
\item{source}{FileSystem URI that contains the source data.}

\item{mountPoint}{The directory of remote source to mount the source.}

\item{extraConfigs}{Extra configurations.}
}
\value{
FALSE to mimic the result if \code{mountPoint} creation fail.
}
\description{
Attach remote storage (Blob, Gen2, Azure File Share) to all working nodes (driver node and worker nodes)
}
\examples{
mssparkutils.fs.mount("abfss://xxx.dfs.core.windows.net", "/mnt")

}
