% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{mssparkutils.fs.put}
\alias{mssparkutils.fs.put}
\title{Writes the given String out to a file, encoded in UTF-8.}
\usage{
mssparkutils.fs.put(file, content, overwrite = FALSE)
}
\arguments{
\item{file}{FileSystem URI.}

\item{content}{Content of file to write, encoded in System default charset.}

\item{overwrite}{If set to TRUE, the file will be overwritten if it existed already.
Note that if overwrite is TRUE and the the write fails, the original file.
may still be deleted.}
}
\value{
FALSE to mimic the result of file put operation fail.
}
\description{
Writes the given String out to a file, encoded in UTF-8.
}
\examples{
mssparkutils.fs.put("/tmp/my-file", "Hello world!", TRUE)

}
