% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{mssparkutils.fs.refreshMounts}
\alias{mssparkutils.fs.refreshMounts}
\title{Refresh workspace level mount points.}
\usage{
mssparkutils.fs.refreshMounts()
}
\value{
FALSE to mimic the refreshMounts fail to refresh mount info.
}
\description{
Refresh workspace level mount points.
}
\examples{
mssparkutils.fs.refreshMounts()

}
