% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{mssparkutils.fs.unmount}
\alias{mssparkutils.fs.unmount}
\title{Removes a mount point.}
\usage{
mssparkutils.fs.unmount(mountPoint)
}
\arguments{
\item{mountPoint}{The directory that was previously mounted.}
}
\value{
FALSE to mimic the result of unmount \code{mountPoint} fail.
}
\description{
Removes a mount point.
}
\examples{
mssparkutils.fs.unmount("/mnt")

}
