% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{mssparkutils.fs.unmountFromDriverNode}
\alias{mssparkutils.fs.unmountFromDriverNode}
\title{Removes a mount point from driver node.}
\usage{
mssparkutils.fs.unmountFromDriverNode(mountPoint)
}
\arguments{
\item{mountPoint}{The directory that was previously mounted.}
}
\value{
TRUE if the mount point was successfully unmounted.
}
\description{
Removes a mount point from driver node.
}
