% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse.R
\name{mssparkutils.lakehouse.create}
\alias{mssparkutils.lakehouse.create}
\title{Create a lakehouse}
\usage{
mssparkutils.lakehouse.create(
  name,
  description = "",
  definition = "",
  workspaceId = ""
)
}
\arguments{
\item{name}{Name of the lakehouse}

\item{description}{Description of the lakehouse}

\item{definition}{Definition of the lakehouse}

\item{workspaceId}{Workspace id of the lakehouse, default to current workspace}
}
\value{
A lakehouse object
}
\description{
Create a lakehouse
}
