% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse.R
\name{mssparkutils.lakehouse.delete}
\alias{mssparkutils.lakehouse.delete}
\title{Delete a lakehouse}
\usage{
mssparkutils.lakehouse.delete(name, workspaceId = "")
}
\arguments{
\item{name}{Name of the lakehouse}

\item{workspaceId}{Workspace id of the lakehouse, default to current workspace}
}
\description{
Delete a lakehouse
}
