% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse.R
\name{mssparkutils.lakehouse.get}
\alias{mssparkutils.lakehouse.get}
\title{Get a lakehouse}
\usage{
mssparkutils.lakehouse.get(name = "", workspaceId = "")
}
\arguments{
\item{name}{Name of the lakehouse}

\item{workspaceId}{Workspace id of the lakehouse, default to current workspace}
}
\value{
A lakehouse object
}
\description{
Get a lakehouse
}
