% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{mssparkutils.notebook.exit}
\alias{mssparkutils.notebook.exit}
\title{This method lets you exit a notebook with a value.}
\usage{
mssparkutils.notebook.exit(value)
}
\arguments{
\item{value}{the value to return when exiting.}
}
\value{
No return value, mimic behavior to set the notebook run exit value using \code{value}.
}
\description{
This method lets you exit a notebook with a value.
}
\examples{
mssparkutils.notebook.exit('exitVal')

}
