% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{mssparkutils.notebook.runMultiple}
\alias{mssparkutils.notebook.runMultiple}
\title{Runs multiple notebooks concurrently with support for dependency relationships.
Details can be found in mssparkutils.notebook.help("runMultiple").}
\usage{
mssparkutils.notebook.runMultiple(pathsOrPipeline)
}
\arguments{
\item{pathsOrPipeline}{A list of notebook names or a complex data structure
(JSON string) that meets the requirements of the
com.microsoft.spark.notebook.msutils.impl.MsNotebookPipeline scala class.}
}
\value{
a list of exit values and exceptions for each notebook
}
\description{
Runs multiple notebooks concurrently with support for dependency relationships.
Details can be found in mssparkutils.notebook.help("runMultiple").
}
