% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{mssparkutils.notebook.updateNBSEndpoint}
\alias{mssparkutils.notebook.updateNBSEndpoint}
\title{provide a way to make people can update the endpoint}
\usage{
mssparkutils.notebook.updateNBSEndpoint(endpoint)
}
\arguments{
\item{endpoint}{the new point}
}
\description{
provide a way to make people can update the endpoint
}
