% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{mssparkutils.runtime.context}
\alias{mssparkutils.runtime.context}
\title{Get runtime properties}
\usage{
mssparkutils.runtime.context()
}
\value{
A dummy env object to mimic the result of runtime context method when used in azure synapse runtime.
}
\description{
Get runtime properties
}
\examples{
mssparkutils.runtime.context()

}
