% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{mssparkutils.runtime.setHcReplId}
\alias{mssparkutils.runtime.setHcReplId}
\title{Set runtime high concurrency mode repl id}
\usage{
mssparkutils.runtime.setHcReplId(replId)
}
\arguments{
\item{replId}{High concurrency mode repl id}
}
\description{
Set runtime high concurrency mode repl id
}
