% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{notebookutils.credentials.getPropertiesAll}
\alias{notebookutils.credentials.getPropertiesAll}
\title{Return all the properties of a given linked service in string format.}
\usage{
notebookutils.credentials.getPropertiesAll(linkedService)
}
\arguments{
\item{linkedService}{Linked service name.}
}
\value{
A empty string used to mimic properties string returned by azure synapse runtime for \code{linkedService}.
}
\description{
Return all the properties of a given linked service in string format.
}
\examples{
notebookutils.credentials.getPropertiesAll('AzureDataLakeStorage1')

}
