% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{notebookutils.credentials.getSecret}
\alias{notebookutils.credentials.getSecret}
\title{Return AKV secret.}
\usage{
notebookutils.credentials.getSecret(akvName, secret, linkedService = NULL)
}
\arguments{
\item{akvName}{Azure Key Vault name.}

\item{secret}{name of the secret being fetched.}

\item{linkedService}{linkedService name of the AKV linked service.}
}
\value{
A empty string used to mimic secret returned by azure synapse runtime for given \code{akvName} and \code{secret}.
}
\description{
Return AKV secret.
}
\examples{
notebookutils.credentials.getSecret('akvName', 'secretName')
notebookutils.credentials.getSecret('akvName', 'secretName', 'AzureDataLakeStorage1')

}
