% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{notebookutils.credentials.getSecretWithLS}
\alias{notebookutils.credentials.getSecretWithLS}
\title{Return AKV secret using linkedService.}
\usage{
notebookutils.credentials.getSecretWithLS(linkedService, secret)
}
\arguments{
\item{linkedService}{linkedService name of the AKV linked service.}

\item{secret}{name of the secret being fetched.}
}
\value{
A empty string used to mimic secret returned by azure synapse runtime for given \code{linkedService} and \code{secret}.
}
\description{
Return AKV secret using linkedService.
}
\examples{
notebookutils.credentials.getSecretWithLS('AzureDataLakeStorage1', 'secretName')

}
