% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{notebookutils.credentials.getToken}
\alias{notebookutils.credentials.getToken}
\title{Get AAD token for a resource.}
\usage{
notebookutils.credentials.getToken(audience, name = "")
}
\arguments{
\item{audience}{token audience.}

\item{name}{token audience.}
}
\value{
A empty string used to mimic token returned by azure synapse runtime for accessing resource \code{audience}.
}
\description{
Get AAD token for a resource.
}
\examples{
notebookutils.credentials.getToken('synapse')
notebookutils.credentials.getToken('storage')
notebookutils.credentials.getToken('storage', 'storage')

}
