% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{notebookutils.credentials.help}
\alias{notebookutils.credentials.help}
\title{Get help message.}
\usage{
notebookutils.credentials.help()
}
\value{
No return value, print empty string to mimic the behavior of help method of mssparkutils credentials module
when used in azure synapse runtime.
}
\description{
Get help message.
}
\examples{
notebookutils.credentials.help()

}
