% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{notebookutils.credentials.putSecretWithLS}
\alias{notebookutils.credentials.putSecretWithLS}
\title{Put AKV secret using linkedService.}
\usage{
notebookutils.credentials.putSecretWithLS(
  linkedService,
  secretName,
  secretValue
)
}
\arguments{
\item{linkedService}{name of AKV linked service.}

\item{secretName}{name of the secret being written.}

\item{secretValue}{value of the secret being written.}
}
\value{
The \code{secretValue} been written.
}
\description{
Put AKV secret using linkedService.
}
\examples{
notebookutils.credentials.putSecretWithLS('AzureDataLakeStorage1', 'secretName', 'secretValue')

}
