% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{notebookutils.env.getClusterId}
\alias{notebookutils.env.getClusterId}
\title{Get cluster id.}
\usage{
notebookutils.env.getClusterId()
}
\value{
A empty string used to mimic cluster id of azure synapse runtime.
}
\description{
Get cluster id.
}
\examples{
notebookutils.env.getClusterId()

}
