% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{notebookutils.env.getPoolName}
\alias{notebookutils.env.getPoolName}
\title{Get pool name.}
\usage{
notebookutils.env.getPoolName()
}
\value{
A empty string used to mimic the name of user's azure synapse spark pool.
}
\description{
Get pool name.
}
\examples{
notebookutils.env.getPoolName()

}
