% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{notebookutils.env.getUserId}
\alias{notebookutils.env.getUserId}
\title{Get user Id.}
\usage{
notebookutils.env.getUserId()
}
\value{
A empty string used to mimic the id of user.
}
\description{
Get user Id.
}
\examples{
notebookutils.env.getUserId()

}
