% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{notebookutils.env.getUserName}
\alias{notebookutils.env.getUserName}
\title{Get user name.}
\usage{
notebookutils.env.getUserName()
}
\value{
A empty string used to mimic the name of user.
}
\description{
Get user name.
}
\examples{
notebookutils.env.getUserName()

}
