% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{notebookutils.env.getWorkspaceName}
\alias{notebookutils.env.getWorkspaceName}
\title{Get workspace name.}
\usage{
notebookutils.env.getWorkspaceName()
}
\value{
A empty string used to mimic the id of the user's azure synapse workspace.
}
\description{
Get workspace name.
}
\examples{
notebookutils.env.getWorkspaceName()

}
