% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{notebookutils.env.help}
\alias{notebookutils.env.help}
\title{Get help message.}
\usage{
notebookutils.env.help()
}
\value{
No return value, print empty string to mimic the behavior of help method of mssparkutils env module
when used in azure synapse runtime.
}
\description{
Get help message.
}
\examples{
notebookutils.env.help()

}
