% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabricClient.R
\name{notebookutils.fabricClient.delete}
\alias{notebookutils.fabricClient.delete}
\title{Send a DELETE request to Fabric.}
\usage{
notebookutils.fabricClient.delete(path, headers = list())
}
\arguments{
\item{path}{Path of the request}

\item{headers}{Headers of the request}
}
\description{
Send a DELETE request to Fabric.
}
