% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabricClient.R
\name{notebookutils.fabricClient.get}
\alias{notebookutils.fabricClient.get}
\title{Send a GET request to Fabric.}
\usage{
notebookutils.fabricClient.get(path, headers = list())
}
\arguments{
\item{path}{Path of the request}

\item{headers}{Headers of the request}
}
\value{
RestResponse Response of the request
}
\description{
Send a GET request to Fabric.
}
