% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabricClient.R
\name{notebookutils.fabricClient.help}
\alias{notebookutils.fabricClient.help}
\title{Get help string for a method.}
\usage{
notebookutils.fabricClient.help(methodName = "")
}
\arguments{
\item{methodName}{Name of the method}
}
\description{
Get help string for a method.
}
