% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabricClient.R
\name{notebookutils.fabricClient.listCapacities}
\alias{notebookutils.fabricClient.listCapacities}
\title{List all capacities in the workspace.}
\usage{
notebookutils.fabricClient.listCapacities(maxResults = 1000L)
}
\arguments{
\item{maxResults}{Maximum number of capacities to return, default is 1000}
}
\value{
Array of Capacity objects
}
\description{
List all capacities in the workspace.
}
