% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabricClient.R
\name{notebookutils.fabricClient.patch}
\alias{notebookutils.fabricClient.patch}
\title{Send a PATCH request to Fabric.}
\usage{
notebookutils.fabricClient.patch(path, content, headers = list())
}
\arguments{
\item{path}{Path of the request}

\item{content}{Content of the request}

\item{headers}{Headers of the request}
}
\value{
RestResponse Response of the request
}
\description{
Send a PATCH request to Fabric.
}
