% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabricClient.R
\name{notebookutils.fabricClient.post}
\alias{notebookutils.fabricClient.post}
\title{Send a POST request to Fabric.}
\usage{
notebookutils.fabricClient.post(path, content, headers = list())
}
\arguments{
\item{path}{Path of the request}

\item{content}{Content of the request}

\item{headers}{Headers of the request}
}
\value{
RestResponse Response of the request
}
\description{
Send a POST request to Fabric.
}
