% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{notebookutils.fs.exists}
\alias{notebookutils.fs.exists}
\title{Check if a file or directory exists.}
\usage{
notebookutils.fs.exists(file)
}
\arguments{
\item{file}{FileSystem URI}
}
\value{
TRUE if the file or directory exists
}
\description{
Check if a file or directory exists.
}
\examples{
\dontrun{
notebookutils.fs.exists("/tmp/my-file")
}

}
