% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{notebookutils.fs.help}
\alias{notebookutils.fs.help}
\title{notebookutils.fs provides utilities for working with various FileSystems.}
\usage{
notebookutils.fs.help(methodName = "")
}
\arguments{
\item{methodName}{method name to get more information.}
}
\value{
No return value, print empty string to mimic the behavior of help method of mssparkutils fs module
when used in azure synapse runtime.
}
\description{
Below is overview about the available methods:
}
\details{
notebookutils.fs.cp: Copies a file or directory, possibly across FileSystems
notebookutils.fs.mv: Moves a file or directory, possibly across FileSystems
notebookutils.fs.ls: Array -> Lists the contents of a directory
notebookutils.fs.mkdirs: Creates the given directory if it does not exist, also creating any necessary parent directories
notebookutils.fs.put: Writes the given String out to a file, encoded in UTF-8
notebookutils.fs.head: Returns up to the first 'maxBytes' bytes of the given file as a String encoded in UTF-8
notebookutils.fs.append: Append the content to a file
notebookutils.fs.rm: Removes a file or directory
}
\examples{
notebookutils.fs.help()
notebookutils.fs.help("ls")

}
