% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{notebookutils.fs.ls}
\alias{notebookutils.fs.ls}
\title{Lists the contents of a directory.}
\usage{
notebookutils.fs.ls(dir)
}
\arguments{
\item{dir}{FileSystem URI}
}
\value{
Empty list to mimic the file list under \code{dir}.
}
\description{
Lists the contents of a directory.
}
\examples{
notebookutils.fs.ls("/tmp/my-folder/")

}
