% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{notebookutils.fs.mkdirs}
\alias{notebookutils.fs.mkdirs}
\title{Creates the given directory if it does not exist, also creating any necessary parent * directories.}
\usage{
notebookutils.fs.mkdirs(dir)
}
\arguments{
\item{dir}{FileSystem URI}
}
\value{
FALSE to mimic the result if \code{dir} creation fail.
}
\description{
Creates the given directory if it does not exist, also creating any necessary parent * directories.
}
\examples{
notebookutils.fs.mkdirs("/tmp/a/b/c")

}
