% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{notebookutils.fs.mounts}
\alias{notebookutils.fs.mounts}
\title{Show information about what is mounted.
Any credentials used to mount the mount points listed will not be displayed.}
\usage{
notebookutils.fs.mounts(extraConfigs = NULL)
}
\arguments{
\item{extraConfigs}{Extra configurations.}
}
\value{
The list of MountPointInfo.
}
\description{
Show information about what is mounted.
Any credentials used to mount the mount points listed will not be displayed.
}
