% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{notebookutils.fs.refreshMounts}
\alias{notebookutils.fs.refreshMounts}
\title{Refresh workspace level mount points.}
\usage{
notebookutils.fs.refreshMounts()
}
\value{
FALSE to mimic the refreshMounts fail to refresh mount info.
}
\description{
Refresh workspace level mount points.
}
\examples{
notebookutils.fs.refreshMounts()

}
