% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{notebookutils.fs.rm}
\alias{notebookutils.fs.rm}
\title{Removes a file or directory.}
\usage{
notebookutils.fs.rm(dir, recurse = FALSE)
}
\arguments{
\item{dir}{FileSystem URI for a single file or a directory.}

\item{recurse}{if TRUE, all files and directories will be recursively deleted.}
}
\value{
FALSE to mimic the result of \code{dir} deletion fail.
}
\description{
Removes a file or directory.
}
\examples{
notebookutils.fs.rm("/tmp/my-folder/", TRUE)

}
