% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{notebookutils.fs.unmount}
\alias{notebookutils.fs.unmount}
\title{Removes a mount point.}
\usage{
notebookutils.fs.unmount(mountPoint)
}
\arguments{
\item{mountPoint}{The directory that was previously mounted.}
}
\value{
FALSE to mimic the result of unmount \code{mountPoint} fail.
}
\description{
Removes a mount point.
}
\examples{
notebookutils.fs.unmount("/mnt")

}
