% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{notebookutils.help}
\alias{notebookutils.help}
\title{Get help message for this module.}
\usage{
notebookutils.help(methodName = "")
}
\arguments{
\item{methodName}{method name to get more information.}
}
\value{
No return value, print empty string to mimic the behavior of help method of mssparkutils module
when used in azure synapse runtime.
}
\description{
Get help message for this module.
}
\examples{
notebookutils.help()

}
