% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse.R
\name{notebookutils.lakehouse.getWithProperties}
\alias{notebookutils.lakehouse.getWithProperties}
\title{Get the info of a Lakehouse with properties.}
\usage{
notebookutils.lakehouse.getWithProperties(name, workspaceId = "")
}
\arguments{
\item{name}{Name of the Lakehouse.}

\item{workspaceId}{Id of the workspace, default to current workspace.}
}
\value{
Artifact object.
Please refer to: https://learn.microsoft.com/en-us/rest/api/fabric
/articles/item-management/properties/lakehouse-properties
}
\description{
Get the info of a Lakehouse with properties.
}
