% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse.R
\name{notebookutils.lakehouse.help}
\alias{notebookutils.lakehouse.help}
\title{The lakehouse module.}
\usage{
notebookutils.lakehouse.help(methodName = "")
}
\arguments{
\item{methodName}{method name to get more information}
}
\description{
notebookutils.lakehouse.create(name: String, description: String, workspaceId: String): Lakehouse -> Create a lakehouse
notebookutils.lakehouse.get(name: String, workspaceId: String): Lakehouse -> Get a lakehouse
notebookutils.lakehouse.delete(name: String, workspaceId: String): void -> Delete a lakehouse
notebookutils.lakehouse.update(name: String, newName: String, description: String, workspaceId: String): Lakehouse -> Update a lakehouse
}
