% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse.R
\name{notebookutils.lakehouse.list}
\alias{notebookutils.lakehouse.list}
\title{List all lakehouses}
\usage{
notebookutils.lakehouse.list(workspaceId = "", maxResults = 1000L)
}
\arguments{
\item{workspaceId}{Workspace id of the lakehouse, default to current workspace}

\item{maxResults}{Maximum number of lakehouses to return, default to 1000}
}
\value{
A list of lakehouse objects
}
\description{
List all lakehouses
}
