% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse.R
\name{notebookutils.lakehouse.listTables}
\alias{notebookutils.lakehouse.listTables}
\title{List all tables in a Lakehouse.}
\usage{
notebookutils.lakehouse.listTables(
  lakehouse = "",
  workspaceId = "",
  maxResults = 1000L
)
}
\arguments{
\item{lakehouse}{Name of the lakehouse}

\item{workspaceId}{Workspace id of the lakehouse, default to current workspace}

\item{maxResults}{Maximum number of tables to return, default to 1000}
}
\value{
A list of table objects
}
\description{
List all tables in a Lakehouse.
}
