% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse.R
\name{notebookutils.lakehouse.loadTable}
\alias{notebookutils.lakehouse.loadTable}
\title{Starts a load table operation.}
\usage{
notebookutils.lakehouse.loadTable(
  loadOption,
  table,
  lakehouse = "",
  workspaceId = ""
)
}
\arguments{
\item{loadOption}{string, loadOption Load options. Please refer to
https://learn.microsoft.com/en-us/rest/api/fabric/lakehouse/tables/load-table}

\item{table}{Name of the table}

\item{lakehouse}{Name of the lakehouse}

\item{workspaceId}{Workspace id of the lakehouse, default to current workspace}
}
\value{
boolean
}
\description{
Starts a load table operation.
}
