% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse.R
\name{notebookutils.lakehouse.update}
\alias{notebookutils.lakehouse.update}
\title{Update a lakehouse}
\usage{
notebookutils.lakehouse.update(
  name,
  newName,
  description = "",
  workspaceId = ""
)
}
\arguments{
\item{name}{Name of the lakehouse}

\item{newName}{New name of the lakehouse}

\item{description}{Description of the lakehouse}

\item{workspaceId}{Workspace id of the lakehouse, default to current workspace}
}
\value{
A lakehouse object
}
\description{
Update a lakehouse
}
