% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakehouse.R
\name{notebookutils.lakehouse.updateDefinition}
\alias{notebookutils.lakehouse.updateDefinition}
\title{Get the definition of a lakehouse}
\usage{
notebookutils.lakehouse.updateDefinition(name, definition, workspaceId = "")
}
\arguments{
\item{name}{Name of the lakehouse}

\item{definition}{Definition of the lakehouse}

\item{workspaceId}{Workspace id of the lakehouse, default to current workspace}
}
\value{
The definition of the lakehouse
}
\description{
Get the definition of a lakehouse
}
