% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebookutils.notebook.create}
\alias{notebookutils.notebook.create}
\title{Create a notebook}
\usage{
notebookutils.notebook.create(
  name,
  description = "",
  content = "",
  defaultLakehouse = "",
  defaultLakehouseWorkspace = "",
  workspaceId = ""
)
}
\arguments{
\item{name}{Name of the notebook}

\item{description}{Description of the notebook}

\item{content}{Definition of the notebook}

\item{defaultLakehouse}{Default lakehouse of the notebook}

\item{defaultLakehouseWorkspace}{Default lakehouse workspace of the notebook}

\item{workspaceId}{Workspace id of the notebook, default to current workspace}
}
\value{
A notebook object
}
\description{
Create a notebook
}
