% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebookutils.notebook.delete}
\alias{notebookutils.notebook.delete}
\title{Delete a notebook}
\usage{
notebookutils.notebook.delete(name, workspaceId = "")
}
\arguments{
\item{name}{Name of the notebook}

\item{workspaceId}{Workspace id of the notebook, default to current workspace}
}
\description{
Delete a notebook
}
