% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebookutils.notebook.get}
\alias{notebookutils.notebook.get}
\title{Get a notebook}
\usage{
notebookutils.notebook.get(name, workspaceId = "")
}
\arguments{
\item{name}{Name of the notebook}

\item{workspaceId}{Workspace id of the notebook, default to current workspace}
}
\value{
A notebook object
}
\description{
Get a notebook
}
