% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebookutils.notebook.help}
\alias{notebookutils.notebook.help}
\title{The notebook module.}
\usage{
notebookutils.notebook.help(methodName = "")
}
\arguments{
\item{methodName}{method name to get more information.}
}
\value{
No return value, print empty string to mimic the behavior of help method of mssparkutils notebook module
when used in azure synapse runtime.
}
\description{
The notebook module.
}
\examples{
notebookutils.notebook.help()
notebookutils.notebook.help("run")

}
