% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebookutils.notebook.list}
\alias{notebookutils.notebook.list}
\title{List all notebooks}
\usage{
notebookutils.notebook.list(workspaceId = "", maxResults = 1000L)
}
\arguments{
\item{workspaceId}{Workspace id of the notebook, default to current workspace}

\item{maxResults}{Maximum number of notebooks to return, default to 1000}
}
\value{
A list of notebook objects
}
\description{
List all notebooks
}
