% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebookutils.notebook.run}
\alias{notebookutils.notebook.run}
\title{Runs a notebook and returns its exit value. The notebook will run
in the current livy session context by default.}
\usage{
notebookutils.notebook.run(path, timeoutSeconds = 90, arguments = NULL)
}
\arguments{
\item{path}{absolute path to the notebook, e.g. /path/to/notebook.}

\item{timeoutSeconds}{timeout in seconds for the called notebook.}

\item{arguments}{string map of arguments to pass to the notebook.}
}
\value{
Empty string to mimic the exitVal set by \code{mssparkutils.notebook.exit}.
}
\description{
Runs a notebook and returns its exit value. The notebook will run
in the current livy session context by default.
}
\examples{
notebookutils.notebook.run('NB1')
notebookutils.notebook.run('NB1', 200)
notebookutils.notebook.run('NB1', 200, list("input"=30))

}
