% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebookutils.notebook.update}
\alias{notebookutils.notebook.update}
\title{Update a notebook}
\usage{
notebookutils.notebook.update(
  name,
  newName,
  description = "",
  workspaceId = ""
)
}
\arguments{
\item{name}{Name of the notebook}

\item{newName}{New name of the notebook}

\item{description}{Description of the notebook}

\item{workspaceId}{Workspace id of the notebook, default to current workspace}
}
\value{
A notebook object
}
\description{
Update a notebook
}
