% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebookutils.notebook.updateDefinition}
\alias{notebookutils.notebook.updateDefinition}
\title{Get the definition of a notebook}
\usage{
notebookutils.notebook.updateDefinition(
  name,
  content,
  defaultLakehouse = "",
  defaultLakehouseWorkspace = "",
  workspaceId = ""
)
}
\arguments{
\item{name}{Name of the notebook}

\item{content}{Definition of the notebook}

\item{defaultLakehouse}{Default lakehouse of the notebook}

\item{defaultLakehouseWorkspace}{Default lakehouse workspace of the notebook}

\item{workspaceId}{Workspace id of the notebook, default to current workspace}
}
\value{
The definition of the notebook
}
\description{
Get the definition of a notebook
}
