% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebookutils.notebook.updateNBSEndpoint}
\alias{notebookutils.notebook.updateNBSEndpoint}
\title{provide a way to make people can update the endpoint}
\usage{
notebookutils.notebook.updateNBSEndpoint(endpoint)
}
\arguments{
\item{endpoint}{the new point}
}
\description{
provide a way to make people can update the endpoint
}
