% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{notebookutils.runtime.help}
\alias{notebookutils.runtime.help}
\title{notebookutils.runtime is a utility to manage runtime context.
context() returns the runtime context as a list.}
\usage{
notebookutils.runtime.help(methodName = "")
}
\arguments{
\item{methodName}{method name to get more information.am}
}
\description{
notebookutils.runtime is a utility to manage runtime context.
context() returns the runtime context as a list.
}
