% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{notebookutils.session.stop}
\alias{notebookutils.session.stop}
\title{Stop an interactive session}
\usage{
notebookutils.session.stop(detach = TRUE)
}
\arguments{
\item{detach}{If detach is True, stop session from standard session,
or detach current notebook from high concurrency session;
if detach is False, stop session in any session. Default is TRUE.}
}
\description{
Stop an interactive session
}
