% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse.R
\name{notebookutils.warehouse.create}
\alias{notebookutils.warehouse.create}
\title{Create a warehouse}
\usage{
notebookutils.warehouse.create(
  name,
  description = "",
  definition = "",
  workspaceId = ""
)
}
\arguments{
\item{name}{Name of the warehouse}

\item{description}{Description of the warehouse}

\item{definition}{Definition of the warehouse}

\item{workspaceId}{Workspace id of the warehouse, default to current workspace}
}
\value{
A warehouse object
}
\description{
Create a warehouse
}
