% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse.R
\name{notebookutils.warehouse.delete}
\alias{notebookutils.warehouse.delete}
\title{Delete a warehouse}
\usage{
notebookutils.warehouse.delete(name, workspaceId = "")
}
\arguments{
\item{name}{Name of the warehouse}

\item{workspaceId}{Workspace id of the warehouse, default to current workspace}
}
\description{
Delete a warehouse
}
