% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse.R
\name{notebookutils.warehouse.get}
\alias{notebookutils.warehouse.get}
\title{Get a warehouse}
\usage{
notebookutils.warehouse.get(name, workspaceId = "")
}
\arguments{
\item{name}{Name of the warehouse}

\item{workspaceId}{Workspace id of the warehouse, default to current workspace}
}
\value{
A warehouse object
}
\description{
Get a warehouse
}
