% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse.R
\name{notebookutils.warehouse.getDefinition}
\alias{notebookutils.warehouse.getDefinition}
\title{Get the definition of a warehouse}
\usage{
notebookutils.warehouse.getDefinition(name, workspaceId = "")
}
\arguments{
\item{name}{Name of the warehouse}

\item{workspaceId}{Workspace id of the warehouse, default to current workspace}
}
\value{
The definition of the warehouse
}
\description{
Get the definition of a warehouse
}
