% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse.R
\name{notebookutils.warehouse.help}
\alias{notebookutils.warehouse.help}
\title{The warehouse module.}
\usage{
notebookutils.warehouse.help(methodName = "")
}
\arguments{
\item{methodName}{method name to get more information}
}
\description{
notebookutils.warehouse.create(name: String, description: String, workspaceId: String): warehouse -> Create a warehouse
notebookutils.warehouse.get(name: String, workspaceId: String): warehouse -> Get a warehouse
notebookutils.warehouse.delete(name: String, workspaceId: String): void -> Delete a warehouse
notebookutils.warehouse.update(name: String, newName: String, description: String, workspaceId: String): warehouse -> Update a warehouse
}
