% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse.R
\name{notebookutils.warehouse.list}
\alias{notebookutils.warehouse.list}
\title{List all warehouses}
\usage{
notebookutils.warehouse.list(workspaceId = "", maxResults = 1000L)
}
\arguments{
\item{workspaceId}{Workspace id of the warehouse, default to current workspace}

\item{maxResults}{Maximum number of warehouses to return, default to 1000}
}
\value{
A list of warehouse objects
}
\description{
List all warehouses
}
