% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{notebookutils.workspace.assignToCapacity}
\alias{notebookutils.workspace.assignToCapacity}
\title{Assign a workspace to a capacity}
\usage{
notebookutils.workspace.assignToCapacity(capacityId, workspaceId = "")
}
\arguments{
\item{capacityId}{Id of the capacity}

\item{workspaceId}{Id of the workspace, default to current workspace}
}
\value{
Boolean indicating success
}
\description{
Assign a workspace to a capacity
}
