% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{notebookutils.workspace.create}
\alias{notebookutils.workspace.create}
\title{Create a workspace}
\usage{
notebookutils.workspace.create(name, description = "", capacityId = "")
}
\arguments{
\item{name}{Name of the workspace}

\item{description}{Description of the workspace}

\item{capacityId}{Id of the capacity, default to current capacity}
}
\value{
A workspace object
}
\description{
Create a workspace
}
