% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{notebookutils.workspace.delete}
\alias{notebookutils.workspace.delete}
\title{Delete a workspace}
\usage{
notebookutils.workspace.delete(workspaceId)
}
\arguments{
\item{workspaceId}{Id of the workspace}
}
\description{
Delete a workspace
}
