% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{notebookutils.workspace.get}
\alias{notebookutils.workspace.get}
\title{Get a workspace}
\usage{
notebookutils.workspace.get(name = "")
}
\arguments{
\item{name}{Name of the workspace}
}
\value{
A workspace object
}
\description{
Get a workspace
}
