% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{notebookutils.workspace.help}
\alias{notebookutils.workspace.help}
\title{The workspace module.}
\usage{
notebookutils.workspace.help(methodName = "")
}
\arguments{
\item{methodName}{method name to get more information}
}
\description{
notebookutils.workspace.assignToCapacity(capacityId: String, workspaceId: String): Boolean -> Assign a workspace to a capacity
notebookutils.workspace.create(name: String, description: String, capacityId: String): workspace -> Create a workspace
notebookutils.workspace.delete(workspaceId: String): void -> Delete a workspace
notebookutils.workspace.get(name: String): workspace -> Get a workspace
notebookutils.workspace.unassignFromCapacity(workspaceId: String): void -> Unassign a workspace from a capacity
notebookutils.workspace.update(workspaceId: String, newName: String, description: String): workspace -> Update a workspace
}
