% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{notebookutils.workspace.list}
\alias{notebookutils.workspace.list}
\title{List all workspaces}
\usage{
notebookutils.workspace.list(maxResults = 1000L)
}
\arguments{
\item{maxResults}{Maximum number of workspaces to return, default to 1000}
}
\value{
A list of workspace objects
}
\description{
List all workspaces
}
