% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{notebookutils.workspace.listArtifacts}
\alias{notebookutils.workspace.listArtifacts}
\title{List the specified artifacts in the workspace}
\usage{
notebookutils.workspace.listArtifacts(
  artifactType,
  workspaceId = "",
  maxResults = 1000L
)
}
\arguments{
\item{artifactType}{Type of the artifact}

\item{workspaceId}{Id of the workspace}

\item{maxResults}{Maximum number of artifacts to return, default to 1000}
}
\value{
A list of artifact objects
}
\description{
List the specified artifacts in the workspace
}
