% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{notebookutils.workspace.unassignFromCapacity}
\alias{notebookutils.workspace.unassignFromCapacity}
\title{Unassign a workspace from a capacity}
\usage{
notebookutils.workspace.unassignFromCapacity(workspaceId)
}
\arguments{
\item{workspaceId}{Id of the workspace}
}
\description{
Unassign a workspace from a capacity
}
