% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{notebookutils.workspace.update}
\alias{notebookutils.workspace.update}
\title{Update a workspace}
\usage{
notebookutils.workspace.update(workspaceId, newName, description = "")
}
\arguments{
\item{workspaceId}{Id of the workspace}

\item{newName}{New name for the workspace}

\item{description}{New description for the workspace}
}
\value{
Updated workspace object
}
\description{
Update a workspace
}
